Update the SE work report template
These changes bring the SE template closer to the guidelines, in terms
of format and structure. Old references (such as the director's name)
have also been updated to be current.

- Use the `blockletter` format
- Add some packages to the preamble
- Formatting changes to the title page and addresses
- Update some of the content
  - Date
  - Current director's name
  - First SE report is submitted after the 2A (second) work term now
- Remove the Contributions section

Note that the template doesn't fully conform to the guidelines. For
example, there is a subsection which contains a single subsubsection;
the guidelines state that a subsection must have zero or at least two
subsubsections.

So basically, it's fine to use the template to get an idea of the
structure of a work report, but you should ultimately refer to the
actual SE work report guidelines.